Fix SOUND_SPEED

+++ main_robot/mvmt/detector.h
@@ -10,7 +10,7 @@ private:
 
 public:
   static constexpr uint8_t TRIGGER_PIN = 23;
-  static constexpr uint8_t SOUND_SPEED = 0.0343;
+  static constexpr float SOUND_SPEED = 0.0343;
   static constexpr uint16_t LOGICAL_THRESHOLD = 2000;
   /* The minimum distance before we consider it a collision. */
   static constexpr float DISTANCE_THRESHOLD = 15.0;
